# Use "jsonv2"/go-json-experimental to marshal OpenAPI v2