--- a/pkg/internal/flags.go
+++ b/pkg/internal/flags.go
@@ -19,3 +19,6 @@ package internal
 // Used by tests to selectively disable experimental JSON unmarshaler
 var UseOptimizedJSONUnmarshaling bool = true
 var UseOptimizedJSONUnmarshalingV3 bool = false
+
+// Used by tests to selectively disable experimental JSON marshaler
+var UseOptimizedJSONMarshaling bool = true