… go.sum
… pkg/validation/spec/schema.go
… pkg/validation/spec/swagger.go
… pkg/validation/spec/swagger_test.go
… test/integration/go.sum
… test/integration/testdata/golden.v2.json