Performance boost is also pretty impressive:
```
name                                    old time/op    new time/op    delta
SwaggerSpec_ExperimentalMarshal/json-8     102ms ± 1%      25ms ± 1%  -75.12%  (p=0.016 n=4+5)

name                                    old alloc/op   new alloc/op   delta
SwaggerSpec_ExperimentalMarshal/json-8    52.4MB ± 4%    19.9MB ± 2%  -62.02%  (p=0.008 n=5+5)

name                                    old allocs/op  new allocs/op  delta
SwaggerSpec_ExperimentalMarshal/json-8      210k ± 0%       76k ± 0%  -63.82%  (p=0.008 n=5+5)
```

Mostly because this removes the need to deserialize json into buffers that are then thrown away when the jsons are concatenated together (due to lots of embedded objects).

@dsnet @alexzielenski @liggitt 

Fixes #324