@apelisse 2022-11-02
Updated with WIP that breaks tests because marshal doesn't serialize in the same order. There's also a slight different change in how unicodes are encoded apparently which is causing the length to not always be strictly equal (though unmarshaling seems to restore to the previous state).

Also, so far I've only found one location in kubernetes/kubernetes that need to be updated: staging/src/k8s.io/apiserver/pkg/util/openapi/proto.go