@dsnet 2022-11-02
If you're not already, you could consider a semantic comparison of JSON objects with:
```go
var got, want []byte = ... // where x and y are raw JSON values
if d := cmp.Diff(got, want, cmp.Transform(in []byte) (out any) {
    // Note that integers larger than 2^53 will lose precision
    // since they will be represented using a Go float64.
    // There are workarounds for this, but calling it out.
    if err := json.Unmarshal(in, &out); err != nil {
        return in
    }
    return out
}); d != "" {
    t.Errorf("JSON mismatch (-got +want):\n%s", d)
}
```