@dsnet 2022-11-02
In v1, the characters in "<>&" were escaped for dubious security reasons. In v2, we stop doing that, and follow RFC recommendations to use the minimal escaped string output.