@apelisse 2022-11-06
Tests should be passing, it's a little bit yucky how the structs are duplicated (more than once), suggestions welcome, ready to merge other than that. Thanks!