@apelisse 2023-02-17
OK, tests pass, should be working now, pretty good improvement considering this is definitely the worst thing going on in an empty api-server right now.